+++ CHANGES.md
@@ -25,6 +25,7 @@
   - ➕ `padStart` / `pad_start`
   - ➕ `repeat`
   - ➕ `replace` => `jsReplace` / `js_replace`
+  - ➕ `replaceAll` / `replace_all`
 
 ## 0.0.1 - 2023-04-28